Custom wallpaper script; April update

+++ scripts/.scripts/facelock.py
@@ -164,14 +164,15 @@ while True:
 
         if match:
             no_match_count = 0
-            p = subprocess.Popen(["ps", "-A", "-o", "pid,etime,comm"], stdout=subprocess.PIPE)
-            out,err = p.communicate()
-            for line in out.decode("utf-8").splitlines():
-                if 'i3lock' in line:
-                    data = line.split(None, 3)
-                    etime = data[1].split(':')
-                    if len(etime) > 2 or int(etime[0]) > 0 or int(etime[1]) > 1:
-                        os.kill(int(data[0]), signal.SIGTERM)
+            if face:
+                p = subprocess.Popen(["ps", "-A", "-o", "pid,etime,comm"], stdout=subprocess.PIPE)
+                out,err = p.communicate()
+                for line in out.decode("utf-8").splitlines():
+                    if 'i3lock' in line:
+                        data = line.split(None, 3)
+                        etime = data[1].split(':')
+                        if len(etime) > 2 or int(etime[0]) > 0 or int(etime[1]) > 1:
+                            os.kill(int(data[0]), signal.SIGTERM)
         else:
             no_match_count += 1
             print(no_match_count)
